pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-audio-transcriber"
version = "0.1.0"
description = "Telegram bot for audio transcription (local faster-whisper + OpenAI fallback)."
requires-python = ">=3.10"
authors = [{ name = "Nataly Project" }]
readme = "README.md"
dependencies = [
  "aiogram>=3.4.1",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "faster-whisper>=1.0.2",
  "openai>=1.45.0",
  "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.8",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "hypothesis>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]


